NATS-queue/requirements.txt
# NATS: Python Asynch IO - https://nats.io/download/nats-io/nats.py/
asyncio-nats-client
